Replaced unsigned int with suitable types

+++ Arrangement_on_surface_2/test/Arrangement_on_surface_2/Point_location_dynamic_test.h
@@ -19,12 +19,12 @@ public:
   typedef typename Base::Point_2                         Point_2;
   typedef typename Base::X_monotone_curve_2              X_monotone_curve_2;
   typedef typename Base::Curve_2                         Curve_2;
-  
+
   typedef typename Base::Arrangement                     Arrangement;
 
   /*! Constructor */
   Point_location_dynamic_test(const Geom_traits& geom_traits);
-  
+
   /*! Destructor */
   virtual ~Point_location_dynamic_test()
   {
@@ -44,20 +44,20 @@ public:
                         curves_filename, queries_filename);
     m_filename_commands.assign(commands_filename);
   }
-  
+
   bool construct_arrangement();
 
 private:
   bool read_perform_opts(std::istream& is);
 
   bool remove(const X_monotone_curve_2& xcv);
-  
+
   /*! The input data file of commands*/
   std::string m_filename_commands;
 };
 
 /*!
- * Constructor. 
+ * Constructor.
  */
 template <typename T_Geom_traits, typename T_Topol_traits>
 Point_location_dynamic_test<T_Geom_traits, T_Topol_traits>::
@@ -81,16 +81,16 @@ construct_arrangement()
     this->print_error(std::string("cannot open file ").append(this->m_filename_commands));
     return false;
   }
-  
+
   if (!read_perform_opts(in_com)) {
     in_com.close();
     return false;
   }
-  in_com.close(); 
-  
+  in_com.close();
+
   // Print the size of the arrangement.
   std::cout << "V = " << this->m_arr->number_of_vertices()
-            << ",  E = " << this->m_arr->number_of_edges() 
+            << ",  E = " << this->m_arr->number_of_edges()
             << ",  F = " << this->m_arr->number_of_faces() << std::endl;
 
   return true;
@@ -103,7 +103,7 @@ read_perform_opts(std::istream& is)
   bool rc = true;
 
   CGAL::Timer timer;
-  timer.reset(); 
+  timer.reset();
   timer.start();
 
   std::string sline;
@@ -117,11 +117,12 @@ read_perform_opts(std::istream& is)
       CGAL::insert(*(this->m_arr),
                    this->m_xcurves.begin(), this->m_xcurves.end());
       // insert(*(this->m_arr), m_points.begin(), m_points.end());
-      CGAL::insert(*(this->m_arr), this->m_curves.begin(), this->m_curves.end());
+      CGAL::insert(*(this->m_arr), this->m_curves.begin(),
+                   this->m_curves.end());
       continue;
     }
-    
-    unsigned int id;
+
+    size_t id;
     line >> id;
     if (id >= this->m_xcurves.size()) {
       std::cerr << "Index of x-monotone curve " << id << " is out of range ("
@@ -137,8 +138,8 @@ read_perform_opts(std::istream& is)
     }
   }
   timer.stop(); ///END
-  std::cout << "Arrangement aggregate construction took " 
-            << timer.time() << std::endl;  
+  std::cout << "Arrangement aggregate construction took "
+            << timer.time() << std::endl;
 
   return rc;
 }
@@ -149,7 +150,7 @@ remove(const X_monotone_curve_2& xcv)
 {
   typedef T_Geom_traits                                 Geom_traits;
   bool rc = false;          // be pasimistic, assume nothing is removed.
-  
+
   const Geom_traits* traits = this->m_arr->geometry_traits();
   typename Geom_traits::Equal_2 equal = traits->equal_2_object();
